+++ sdwan_feature_profiles.tf
@@ -1,3 +1,9 @@
+resource "sdwan_application_priority_feature_profile" "application_priority_feature_profile" {
+  for_each    = { for t in try(local.feature_profiles.application_priority_profiles, {}) : t.name => t }
+  name        = each.value.name
+  description = try(each.value.description, "")
+}
+
 resource "sdwan_cli_feature_profile" "cli_feature_profile" {
   for_each    = { for t in try(local.feature_profiles.cli_profiles, {}) : t.name => t }
   name        = each.value.name