+++ sdwan_policy_objects.tf
@@ -34,7 +34,7 @@ resource "sdwan_policy_object_as_path_list" "policy_object_as_path_list" {
 }
 
 resource "sdwan_policy_object_class_map" "policy_object_class_map" {
-  for_each           = { for p in try(local.feature_profiles.policy_object_profile.class_maps, {}) : p.name => p }
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.forwarding_classes, {}) : p.name => p }
   name               = each.value.name
   description        = try(each.value.description, null)
   feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id